requirements.txt
beanie==1.26.0
motor==3.5.1
paho-mqtt==2.1.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-miio==0.6.0.dev0
APScheduler==3.10.4